Add a sirius-legacy repo
It will be used to migrate the contents of
ssh://git.eclipse.org:29418/sirius/org.eclipse.sirius.legacy. The code
there is no longer maintained, but we'd like not to completely lose it
when Gerrit will be closed.

The GitHub repo will probably be moved into an "archived" state once
I've pushed this historical code in it.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>